+++ src/EndpointInterface.php
@@ -21,7 +21,9 @@ interface EndpointInterface
 
     const PATH_COUNTRIES = '/auth/countries';
     const PATH_BANKS = '/auth/banks';
+    const PATH_PAYMENT_METHODS = '/auth/paymentMethods';
     const PATH_BANK = '/auth/banks/{bankId}';
+    const PATH_BANK_BY_CARD_NUMBER_PIECE = '/auth/banks/cards/{cardNumberPiece}';
     const PATH_AUTH = '/auth';
     const PATH_RECEIVE_TOKEN = '/auth/token';
     const PATH_REFRESH_TOKEN = '/auth/token';